+++ src/app/test-results/test-results-list/test-results-detail-list-item.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from '@angular/core';
+import { ResponseTestResultDetail } from 'api-client';
+
+@Component({
+  selector: 'wh-test-results-detail-list-item',
+  templateUrl: './test-results-detail-list-item.component.html',
+})
+export class TestResultsDetailsListItemComponent {
+  @Input() detail: ResponseTestResultDetail;
+  visible: boolean;
+
+  constructor() { }
+}