+++ src/app/test-results/test-results-list/test-results-summary-list-item.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+import { ResponseTestResultSummary } from 'api-client';
+
+@Component({
+  selector: 'wh-test-results-summary-list-item',
+  templateUrl: './test-results-summary-list-item.component.html',
+})
+export class TestResultsSummaryListItemComponent {
+  @Input() summary: ResponseTestResultSummary;
+
+  constructor() { }
+}