+++ src/app/test-results/test-results-list/test-results-summary-list.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+import { ResponseBuild } from 'api-client';
+
+@Component({
+  selector: 'wh-test-results-summary-list',
+  templateUrl: './test-results-summary-list.component.html',
+})
+export class TestResultsSummaryListComponent {
+  @Input() build: ResponseBuild;
+
+  constructor() { }
+}