+++ src/app/models/main-project.model.ts
@@ -3,18 +3,18 @@ import { ProjectBuild } from './project-build.model';
 import { ProjectAction } from './project-action.model';
 
 export class WharfProject implements ResponseProject {
-    avatarUrl?: string;
-    branches?: ResponseBranch[];
-    buildDefinition?: string;
-    buildHistory?: ResponseBuild[];
-    description?: string;
-    gitUrl?: string;
-    groupName?: string;
-    name?: string;
-    projectId?: number;
-    provider?: ResponseProvider;
-    providerId?: number;
-    tokenId?: number;
-    build?: ProjectBuild;
-    actions?: ProjectAction[];
+  avatarUrl?: string;
+  branches?: ResponseBranch[];
+  buildDefinition?: string;
+  buildHistory?: ResponseBuild[];
+  description?: string;
+  gitUrl?: string;
+  groupName?: string;
+  name?: string;
+  projectId?: number;
+  provider?: ResponseProvider;
+  providerId?: number;
+  tokenId?: number;
+  build?: ProjectBuild;
+  actions?: ProjectAction[];
 }