+++ CHANGELOG.md
@@ -26,6 +26,12 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
   - `wharf-provider-github` from v2.0.0-rc to v2.0.0.
   - `wharf-provider-gitlab` from v1.2.0-rc to v1.3.0.
 
+- Added test results failed/skipped/passed numbers to each build in a project's
+  build list. (#53)
+
+- Added test result details, showing test messages and per-artifact test
+  results, to the build details page. (#53)
+
 - Security: Changed version of `wharf-collect-licenses` from v1.0.0 to v2.0.0,
   which includes upgrade of `node-fetch` from v2.6.5 to v3.2.0, fixing
   [CVE-2022-0235](https://nvd.nist.gov/vuln/detail/CVE-2022-0235). (#110)